Split the 'test_helpers' file.
