+++ src/test_helpers.rs
@@ -1,49 +1,10 @@
 // Copyright (c) ZeroC, Inc.
 
-use crate::ast::Ast;
+//! This module contains helper functions that are useful for testing both slicec and the compilers that use it.
+//! For the test helpers that are specific to slicec (and hence not exported, see: 'tests/slicec_test_helpers.rs').
+
 use crate::compilation_state::CompilationState;
-use crate::compile_from_strings;
 use crate::diagnostics::Diagnostic;
-use crate::slice_options::SliceOptions;
-
-/// This function parses the provided Slice file.
-/// It is the lowest level test helper function, returning a full [`CompilationState`] instead of only part of it.
-/// It also allows tests to configure the compiler by passing in [`SliceOptions`].
-#[must_use]
-pub fn parse(slice: impl Into<String>, options: Option<SliceOptions>) -> CompilationState {
-    compile_from_strings(&[&slice.into()], options, |_| {}, |_| {})
-}
-
-/// This function parses the provided Slice file and returns the AST generated by doing so.
-/// If any errors are encountered during parsing, it panics.
-#[must_use]
-pub fn parse_for_ast(slice: impl Into<String>) -> Ast {
-    let compilation_state = parse(slice, None);
-    if compilation_state.diagnostic_reporter.has_errors() { // TODO should this also check warnings?
-        panic!("{:?}", compilation_state.diagnostic_reporter);
-    }
-    compilation_state.ast
-}
-
-/// This function parses the provided Slice file and returns any Diagnostics that were emitted during parsing.
-#[must_use]
-pub fn parse_for_diagnostics(slice: impl Into<String>) -> Vec<Diagnostic> {
-    parse_multiple_for_diagnostics(&[&slice.into()])
-}
-
-/// This function parses the provided Slice files and returns any Diagnostics that were emitted during parsing.
-/// Each string is treated as a separate Slice file by the parser.
-#[must_use]
-pub fn parse_multiple_for_diagnostics(slice: &[&str]) -> Vec<Diagnostic> {
-    diagnostics_from_compilation_state(compile_from_strings(slice, None, |_| {}, |_| {}))
-}
-
-/// Asserts that the provided slice parses okay, producing no errors.
-pub fn assert_parses(slice: impl Into<String>) {
-    let diagnostics = parse_for_diagnostics(slice);
-    let expected: [Diagnostic; 0] = []; // Compiler needs the type hint.
-    check_diagnostics(diagnostics, expected);
-}
 
 /// This function is used to get the Diagnostics from a `CompilationState`.
 #[must_use]
