+++ tests/class_tests.rs
@@ -1,3 +1,5 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
 mod classes;
