+++ tests/classes/container.rs
@@ -1,8 +1,8 @@
 // Copyright (c) ZeroC, Inc.
 
+use crate::test_helpers::*;
 use slice::diagnostics::{Diagnostic, Error};
 use slice::grammar::*;
-use slice::test_helpers::*;
 use test_case::test_case;
 
 /// Verifies that classes can contain fields.
