+++ tests/classes/encoding.rs
@@ -2,9 +2,9 @@
 
 mod slice2 {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::Encoding;
-    use slice::test_helpers::*;
 
     #[test]
     fn unsupported_error() {
