+++ tests/classes/inheritance.rs
@@ -1,9 +1,9 @@
 // Copyright (c) ZeroC, Inc.
 
+use crate::test_helpers::*;
 use slice::diagnostics::{Diagnostic, Error};
 use slice::grammar::*;
 use slice::slice_file::Span;
-use slice::test_helpers::*;
 
 #[test]
 fn supports_single_inheritance() {
