+++ tests/comment_tests.rs
@@ -1,10 +1,12 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
 mod comments {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error, Warning};
     use slice::grammar::*;
-    use slice::test_helpers::*;
     use test_case::test_case;
 
     #[test]
