+++ tests/encoding_tests.rs
@@ -1,9 +1,11 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
 mod encodings {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
-    use slice::test_helpers::*;
     use test_case::test_case;
 
     /// Verifies that the supported encodings compile
