+++ tests/enums/encoding.rs
@@ -2,9 +2,9 @@
 
 mod slice1 {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::Encoding;
-    use slice::test_helpers::*;
 
     /// Verifies that the slice parser with the Slice1 encoding emits errors when parsing an enum
     /// that has an underlying type.
@@ -39,8 +39,8 @@ mod slice1 {
 
 mod slice2 {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
-    use slice::test_helpers::*;
     use test_case::test_case;
 
     #[test_case("uint8"; "uint8")]
