+++ tests/module_tests.rs
@@ -1,9 +1,11 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
 mod module {
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::*;
-    use slice::test_helpers::*;
     use test_case::test_case;
 
     #[test_case("{}", false; "normal")]
