+++ tests/parser_tests.rs
@@ -1,8 +1,10 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
+use crate::test_helpers::*;
 use slice::diagnostics::{Diagnostic, Error};
 use slice::slice_file::Span;
-use slice::test_helpers::*;
 
 #[test]
 fn parse_empty_string() {
