+++ tests/primitives/mod.rs
@@ -2,8 +2,8 @@
 
 mod encoding;
 
+use crate::test_helpers::*;
 use slice::grammar::*;
-use slice::test_helpers::parse_for_ast;
 use test_case::test_case;
 
 #[test_case("bool", Primitive::Bool, 2; "bool")]
