+++ tests/scope_resolution_tests.rs
@@ -1,10 +1,12 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
 mod scope_resolution {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::*;
-    use slice::test_helpers::*;
 
     #[test]
     fn file_scoped_modules_can_not_contain_sub_modules() {
