+++ tests/structs/encoding.rs
@@ -2,9 +2,9 @@
 
 mod slice1 {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::Encoding;
-    use slice::test_helpers::*;
 
     /// Verifies using the slice parser with Slice1 will emit errors when parsing
     /// non-compact structs.
@@ -36,9 +36,9 @@ mod slice1 {
 
 mod slice2 {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::Encoding;
-    use slice::test_helpers::*;
 
     /// Verifies using the slice parser with Slice2 will emit errors when parsing
     /// structs that contain Slice1 types.
