+++ tests/tag_tests.rs
@@ -1,10 +1,12 @@
 // Copyright (c) ZeroC, Inc.
 
+mod test_helpers;
+
 mod tags {
 
+    use crate::test_helpers::*;
     use slice::diagnostics::{Diagnostic, Error};
     use slice::grammar::*;
-    use slice::test_helpers::*;
     use test_case::test_case;
 
     #[test]
